Add provider-less backend configuration tests
- Introduced tests for provider-less configurations in store backends, validating the presence of connection strings for PostgreSQL, ClickHouse, and remote blob storage.
- Ensured that the tests assert correct behavior when connection strings are absent, confirming the expected defaults.
- Broke down store_config_tests.rs to appropriate sub-tests using an idiomatic Rust approach.
- These tests are designed to fail against the current codebase, highlighting gaps in the implementation as per specification §5.1, CLI-1480.
