+++ bitloops/src/config/store_config_tests/events.rs
@@ -1,0 +1,71 @@
+use super::*;
+
+#[test]
+fn events_backend_duckdb_path_defaults_under_repo_store_directory() {
+    let events = EventsBackendConfig {
+        provider: EventsProvider::DuckDb,
+        duckdb_path: None,
+        clickhouse_url: None,
+        clickhouse_user: None,
+        clickhouse_password: None,
+        clickhouse_database: None,
+    };
+
+    let resolved = events.duckdb_path_or_default();
+    let rendered = resolved.to_string_lossy();
+    assert!(
+        rendered.ends_with(".bitloops/stores/event/events.duckdb")
+            || rendered.ends_with(".bitloops\\stores\\event\\events.duckdb")
+    );
+}
+
+#[test]
+fn events_backend_duckdb_path_preserves_explicit_path() {
+    let events = EventsBackendConfig {
+        provider: EventsProvider::DuckDb,
+        duckdb_path: Some("/tmp/custom-events.duckdb".to_string()),
+        clickhouse_url: None,
+        clickhouse_user: None,
+        clickhouse_password: None,
+        clickhouse_database: None,
+    };
+
+    assert_eq!(
+        events.duckdb_path_or_default(),
+        PathBuf::from("/tmp/custom-events.duckdb")
+    );
+}
+
+#[test]
+fn events_backend_clickhouse_endpoint_uses_defaults() {
+    let events = EventsBackendConfig {
+        provider: EventsProvider::ClickHouse,
+        duckdb_path: None,
+        clickhouse_url: None,
+        clickhouse_user: None,
+        clickhouse_password: None,
+        clickhouse_database: None,
+    };
+
+    assert_eq!(
+        events.clickhouse_endpoint(),
+        "http://localhost:8123/?database=default"
+    );
+}
+
+#[test]
+fn events_backend_clickhouse_endpoint_trims_trailing_slash() {
+    let events = EventsBackendConfig {
+        provider: EventsProvider::ClickHouse,
+        duckdb_path: None,
+        clickhouse_url: Some("http://localhost:8123/".to_string()),
+        clickhouse_user: None,
+        clickhouse_password: None,
+        clickhouse_database: Some("bitloops".to_string()),
+    };
+
+    assert_eq!(
+        events.clickhouse_endpoint(),
+        "http://localhost:8123/?database=bitloops"
+    );
+}
